pyproject.toml
# Initially generated by nens-meta

[build-system]
requires = ["setuptools>=69"]

[project]
name = "brostar-api"
requires-python = ">=3.11"
dependencies = [
    "celery[redis]",
    "django-cors-headers",
    "django-encrypted-model-fields",
    "django-filter",
    "django==5.0.*",
    "djangorestframework",
    "djangorestframework-api-key",
    "drf-yasg",
    "markdown > 3",  # DRF api docs
    "mkdocs-material",
    "nens-auth-client",
    "openpyxl",
    "pandas",
    "polars",
    "psycopg2-binary",
    "pydantic",
    "pytest-django",
    "pytest-mock",
    "requests",
    "sentry-sdk[django]",
    "whitenoise",
    "xmltodict",
]
description = "Simplify data management for BRO"
authors = []
license = {file = "LICENSE.txt"}
readme = "README.md"
version = "0.78.dev0"

[project.optional-dependencies]
test = ["pytest"] # pytest added by nens-meta

[tool.setuptools]
packages = ["brostar_api", "api", "gmn", "gmw", "gld", "gar", "frd"]

[tool.pytest.ini_options]
testpaths = ["brostar_api", "api", "gmn", "gmw", "gld", "gar", "frd"]
log_level = "DEBUG"
django_find_project = false
DJANGO_SETTINGS_MODULE = "brostar_api.settings"

[tool.coverage.run]
source = ["brostar_api", "api", "gmn", "gmw", "gld", "gar", "frd"]

[tool.coverage.report]
show_missing = true
skip_empty = true

[tool.ruff]
target-version = "py311"

[tool.ruff.lint]
select = ["E4", "E7", "E9", "F", "I", "UP", "C901"]

[tool.zest-releaser]
release = false

[tool.mypy]
python_version = "3.12"
check_untyped_defs = true
disallow_untyped_defs = true
ignore_missing_imports = true
exclude = ["migrations/*", "tests/*"]


[tool.dependencychecker]
djangorestframework = ["rest_framework"]
django-cors-headers = ["corsheaders"]
django-filter = ["django_filters"]
django-encrypted-model-fields = ["encrypted_model_fields"]
ignore-packages = ["api", "gmn", "gmw", "gld", "gar", "frd", "psycopg2-binary", "markdown", "pytest-django"]

[tool.pyright]
include = ["brostar_api", "api", "gmn", "gmw", "gld", "gar", "frd"]
